requirements.txt
pandas==2.2.2
requests==2.31.0
python-docx==1.1.0
python-dotenv==1.0.1
python-pptx==0.6.23
PyPDF2==3.0.1
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-text-splitters==0.0.1
langsmith==0.1.51
O365 # for demo1.py